% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipSKILL.R
\name{zipSKILL}
\alias{zipSKILL}
\title{To create a \link[utils]{zip} SKILL File for Claude}
\usage{
zipSKILL(
  from = ".",
  to = "~/Documents",
  flags = "-q",
  extras = "-x _ignore/*",
  ...
)
}
\arguments{
\item{from}{\link[base]{character} scalar, input directory, default value is the working directory \code{.}.}

\item{to}{\link[base]{character} scalar, output directory, default value is the \verb{~/Documents} directory.}

\item{flags, extras, ...}{additional parameters of the function \link[utils]{zip}}
}
\description{
To create a \link[utils]{zip} SKILL File for Claude
}
