% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipSKILL_github.R
\name{zipSKILL_github}
\alias{zipSKILL_github}
\title{To create a \link[utils]{zip} SKILL File for Claude from GitHub}
\usage{
zipSKILL_github(repo, ...)
}
\arguments{
\item{repo}{\link[base]{character} scalar, in the format of \code{'[user]/[repo]'}}

\item{...}{parameters of the function \code{\link[=zipSKILL]{zipSKILL()}}, except for \code{from}}
}
\description{
To create a \link[utils]{zip} SKILL File for Claude from GitHub
}
\examples{
\dontrun{
zipSKILL_github('tingtingzhan/factcheck')
}

}
