% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipSkill.R
\name{zipSKILL}
\alias{zipSKILL}
\title{To create a \link[utils]{zip} File with Only \code{.md} files}
\usage{
zipSKILL(from = ".", to = "~/document")
}
\arguments{
\item{from}{\link[base]{character} scalar, input directory,
default value is the working directory \code{.}.}

\item{to}{\link[base]{character} scalar, input directory,
default value is the \verb{~/document} directory.}
}
\description{
To create a \link[utils]{zip} File with Only \code{.md} files
}
