% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipSkill.R
\name{zipSKILL}
\alias{zipSKILL}
\title{To create a \link[utils]{zip} File with Only \code{.md} files}
\usage{
zipSKILL(path = ".")
}
\arguments{
\item{path}{\link[base]{character} scalar}
}
\description{
To create a \link[utils]{zip} File with Only \code{.md} files
}
