% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipSKILL_github.R
\name{zipSKILL_github}
\alias{zipSKILL_github}
\title{To create a \link[utils]{zip} File from GitHub}
\usage{
zipSKILL_github(repo, ...)
}
\arguments{
\item{repo}{..}

\item{...}{..}
}
\description{
To create a \link[utils]{zip} File from GitHub
}
\examples{
\dontrun{
zipSKILL_github('tingtingzhan/factcheck')
}

}
